+++ mininterface/types/internal.py
@@ -1,6 +1,5 @@
 RecommendedWidget = type('BoolWidget', (), {})
 BoolWidget = type('BoolWidget', (RecommendedWidget,), {})
-EnumWidget = type('EnumWidget', (RecommendedWidget,), {})
 CallbackButtonWidget = type('CallableWidget', (RecommendedWidget,), {})
 SubmitButtonWidget = type('SubmitButtonWidget', (RecommendedWidget,), {})
 """ NOTE EXPERIMENTAL Special type: Submit button """