+++ tests/configs.py
@@ -10,6 +10,7 @@ from mininterface import Tag
 from mininterface.subcommands import Command
 from mininterface.types import CallbackTag, PathTag
 from mininterface.types.alias import Choices, Validation
+from mininterface.types.rich_tags import EnumTag
 from mininterface.validators import not_empty
 
 
@@ -146,7 +147,7 @@ class ConstrainedEnv:
 
     test2: Annotated[str, Validation(not_empty)] = "hello"
 
-    choices: Annotated[str, Choices("one", "two")] = "one"
+    choices: Annotated[str, EnumTag(choices=("one", "two"))] = "one"
 
 
 @dataclass